Move getActiveSnapshotEdges to selectors/index.js
